requirements.txt
# Core PDF processing
PyMuPDF>=1.23.0
pdf2image>=1.16.0
Pillow>=10.0.0

# CLI and rich interface
click>=8.1.0
rich>=13.0.0
typer>=0.9.0

# PDF generation and manipulation
reportlab>=4.0.0
pdfrw>=0.4

# OCR capabilities
pytesseract>=0.3.10

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.5.0